requirements.txt
absl-py==1.0.0
apturl==0.5.2
asn1crypto==0.24.0
astunparse==1.6.3
Brlapi==0.6.6
cachetools==5.0.0
certifi==2018.1.18
chardet==3.0.4
charset-normalizer==2.0.11
click==6.7
cmake==3.22.2
collection==0.1.6
colorama==0.3.7
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
cycler==0.11.0
defer==1.0.6
distro-info===0.18ubuntu0.18.04.1
elasticsearch==7.16.1
face-recognition-models==0.3.0
facenet==1.0.5
flatbuffers==2.0
fonttools==4.29.1
gast==0.5.3
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.43.0
h5py==3.6.0
httplib2==0.9.2
idna==2.6
image-match @ file:///media/arg/Professional/PYTHON/videocheck/codes/image-match
imageio==2.13.5
importlib-metadata==4.10.1
imutils==0.5.4
joblib==1.1.0
keras==2.8.0
Keras-Preprocessing==1.1.2
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.3.2
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
libclang==13.0.0
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.3.6
MarkupSafe==1.0
matplotlib==3.5.1
mtcnn==0.1.1
netifaces==0.10.4
networkx==2.6.3
numpy==1.21.5
oauth==1.0.1
oauthlib==3.2.0
olefile==0.45.1
opencv-contrib-python==4.5.5.62
opencv-python==4.5.5.62
opt-einsum==3.3.0
packaging==21.3
pexpect==4.2.1
Pillow==8.4.0
protobuf==3.19.4
psutil==5.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
PyGObject==3.26.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==3.0.6
pyRFC3339==1.0
python-apt==1.6.5+ubuntu0.7
python-dateutil==2.8.2
python-debian==0.1.32
pytz==2018.3
PyWavelets==1.2.0
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.27.1
requests-oauthlib==1.3.1
requests-unixsocket==0.1.5
rsa==4.8
scikit-image==0.19.1
scikit-learn==1.0.2
scipy==1.7.3
SecretStorage==2.3.1
simplejson==3.13.2
six==1.16.0
system-service==0.3
systemd-python==234
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
threaded==4.1.0
threadpoolctl==3.1.0
tifffile==2021.11.2
tk==0.1.0
typing_extensions==4.0.1
ubuntu-advantage-tools==27.4
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.22
usb-creator==0.3.3
virtualenv==15.1.0
wadllib==1.3.2
Werkzeug==2.0.2
wrapt==1.13.3
xkit==0.0.0
zipp==3.7.0
zope.interface==4.3.2
